+++ packages/guide/src/app/database/database.component.ts
@@ -7,6 +7,7 @@ import { FormControl } from "@angular/forms";
 import { concat, debounceTime, map, Observable, take, tap } from "rxjs";
 import { DatabaseDetailsComponent } from "./components/database-details/database-details.component";
 import { coerceBooleanProperty } from "@angular/cdk/coercion";
+import { Title } from "@angular/platform-browser";
 
 @Component({
     selector: 'app-database',
@@ -25,19 +26,20 @@ export class DatabaseComponent {
     private _didInitialLoad = false;
 
     constructor(
-        private route: ActivatedRoute,
-        private router: Router,
-        private database: DatabaseService,
-        private appRef: ApplicationRef,
-        private injector: EnvironmentInjector
+        private _route: ActivatedRoute,
+        private _router: Router,
+        private _database: DatabaseService,
+        private _appRef: ApplicationRef,
+        private _injector: EnvironmentInjector,
+        private _title: Title,
     ) {
-        this.database.getDatabaseDetails().pipe(
+        this._database.getDatabaseDetails().pipe(
             tap(() => this.allPrefetched = true),
             take(1)
         ).subscribe();
         this.shouldHideImportantNote = coerceBooleanProperty(localStorage.getItem(this._localStorageHideNoteKey));
 
-        this.items = database.getItems().filter(item => getQuality(item.id) === Quality.BASE);
+        this.items = _database.getItems().filter(item => getQuality(item.id) === Quality.BASE);
         const mapToItems = map<string, Item[]>(searchTerm => {
             const searchString = searchTerm.toLocaleLowerCase();
             const items = this.items.filter(item => item.displayName.toLocaleLowerCase().includes(searchString) || (searchString.startsWith('item_') && item.id.startsWith(searchString)));
@@ -48,7 +50,7 @@ export class DatabaseComponent {
 
         this.filteredItems$ =
             concat(
-                this.route.queryParams.pipe(map(params => {
+                this._route.queryParams.pipe(map(params => {
                         const value = params['q'] ?? '';
                         this.searchTermControl.setValue(value, {emitEvent: false});
                         return value;
@@ -101,13 +103,13 @@ export class DatabaseComponent {
 
     insertAfter(newNode: ComponentRef<DatabaseDetailsComponent>, existingNode: HTMLElement) {
         existingNode.parentNode?.insertBefore(newNode.location.nativeElement, existingNode.nextSibling);
-        this.appRef.attachView(newNode.hostView);
+        this._appRef.attachView(newNode.hostView);
     }
 
 
     createComponent(value: Item): ComponentRef<DatabaseDetailsComponent> {
         const componentRef = createComponent(DatabaseDetailsComponent, {
-            environmentInjector: this.injector
+            environmentInjector: this._injector
         });
 
         componentRef.setInput('item', value)
@@ -128,10 +130,10 @@ export class DatabaseComponent {
     public updateQueryParam(searchTerm: string) {
 
 
-        this.router.navigate(
+        this._router.navigate(
             [],
             {
-                relativeTo: this.route,
+                relativeTo: this._route,
                 queryParams: {q: searchTerm},
                 queryParamsHandling: 'merge',
                 replaceUrl: true
@@ -140,18 +142,18 @@ export class DatabaseComponent {
 
     public updateRouteParam(itemId: string) {
 
-        this.router.navigate(
+        this._router.navigate(
             ['..', itemId],
             {
-                relativeTo: this.route,
+                relativeTo: this._route,
                 replaceUrl: true,
                 queryParamsHandling: "preserve"
-            });
+            }).then(() => !!this.selectedItem && this.updateTitle(this.selectedItem.displayName));
     }
 
     initialItemLoad(): void {
         if (this._didInitialLoad) return;
-        this.route.params.pipe(
+        this._route.params.pipe(
             tap(params => {
                 const itemId = params['itemId'];
                 this._didInitialLoad = true;
@@ -165,5 +167,12 @@ export class DatabaseComponent {
             }),
             take(1)
         ).subscribe()
+    }
+
+    protected updateTitle(itemName: string) {
+        const title = this._title.getTitle();
+        if (title) {
+            this._title.setTitle(`${itemName} - ${title}`)
+        }
     }
 }