+++ packages/guide/src/app/journal/components/base-journal-page/base-journal-page.component.ts
@@ -1,12 +1,12 @@
 import { ChangeDetectorRef, Component, inject, ViewChild } from '@angular/core';
 import { BaseCrop, Item, JournalOrder, Season } from '@ci/data-types';
 import { combineLatest, map, Observable, of, startWith } from 'rxjs';
-import { DatabaseService } from '../../../shared/services/database.service';
 import { UiIcon } from '../../../shared/enums/ui-icon.enum';
 import { MediaMatcher } from '@angular/cdk/layout';
 import { FormGroup } from '@angular/forms';
 import { FilterForm } from "../../../shared/types/filter-form.type";
 import { MatTabGroup } from "@angular/material/tabs";
+import { BaseTabbedSelectableContainerComponent } from "../../../shared/components/base-tabbed-selectable-container/base-tabbed-selectable-container.component";
 
 export interface BaseJournalPageComponent<D> {
     filterPredicate?(foundEntry: D, filterValues: BaseJournalPageComponent<D>['formControl']['value'], index: number): boolean;
@@ -16,24 +16,22 @@ export interface BaseJournalPageComponent<D> {
 @Component({
     template: ''
 })
-export class BaseJournalPageComponent<D extends ({ item: Item } | Item)> {
+export class BaseJournalPageComponent<D extends ({ item: Item } | Item)> extends BaseTabbedSelectableContainerComponent<D> {
 
     @ViewChild(MatTabGroup) matTabGroup?: MatTabGroup
 
     uiIcon = UiIcon;
     tabs: { title: string; data: Observable<D[]> }[] = [];
-    openDrawer = false;
-    selectedEntity?: D;
+
     season = Season;
     formControl: FormGroup<FilterForm>;
     mobileQuery: MediaQueryList;
     media: MediaMatcher = inject(MediaMatcher);
-    protected showTable = false;
-    protected readonly _database: DatabaseService = inject(DatabaseService);
     private changeDetectorRef: ChangeDetectorRef = inject(ChangeDetectorRef);
     private readonly _mobileQueryListener: () => void;
 
     constructor(formControl: FormGroup<FilterForm>) {
+        super();
         this.mobileQuery = this.media.matchMedia('(max-width: 600px)');
         this._mobileQueryListener = () => this.changeDetectorRef.detectChanges();
         this.mobileQuery.addListener(this._mobileQueryListener);
@@ -76,11 +74,6 @@ export class BaseJournalPageComponent<D extends ({ item: Item } | Item)> {
                 return res;
             })
         );
-    }
-
-    showDetails(fishEntry?: D) {
-        this.selectedEntity = fishEntry;
-        this.openDrawer = true;
     }
 
     ngOnDestroy(): void {