+++ packages/guide/src/app/locations/components/lake-temple/lake-temple.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { MinimalItem, OfferingAltar, Offerings } from "@ci/data-types";
-import { Observable, take, tap } from "rxjs";
+import { Observable, tap } from "rxjs";
 import { BaseTabbedSelectableContainerComponent } from "../../../shared/components/base-tabbed-selectable-container/base-tabbed-selectable-container.component";
 
 @Component({
@@ -16,18 +16,7 @@ export class LakeTempleComponent extends BaseTabbedSelectableContainerComponent<
         this.offerings$ = this._database.fetchOfferings$().pipe(
             tap((records) => {
                     const altarNames = records.map(altar => altar.offeringGroupTitle);
-                    this._route.paramMap.pipe(
-                        tap(params => {
-
-                            const altar = params.get('altar');
-
-                            if (altar)
-                                this.selectedTabIndex = altarNames.map(s => s.toLowerCase().replaceAll(' ', '')).indexOf(altar);
-                        }),
-                        take(1)
-                    ).subscribe();
-
-
+                this.activateTabFromRoute(altarNames);
                 }
             )
         );