+++ packages/guide/CHANGELOG.md
@@ -4,10 +4,12 @@
 
 #### Added
 
+- Added direct urls to journal tabs
+- Added title to views so that the browser tab name should be more useful
 - Added quality prices to database if existent
 - Added locations
     - Added lake temple with offerings as first location
-      - Added offerings to database
+        - Added offerings to database
 - Added data tables to all journal and crafting pages
 - Added sorting to existing tables (note: not all columns are sortable)
 - Added highlight for selected item in grid view